requirements.txt
pandas
psycopg2
python-dotenv
sqlalchemy
mysql-connector
mysqlclient
requests